requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
cffi==1.16.0
cryptography==41.0.7
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
ecdsa==0.18.0
graphene==3.3
graphene-django==3.1.5
graphql-core==3.2.3
graphql-relay==3.2.0
promise==2.3
psycopg2==2.9.9
pyasn1==0.5.1
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2023.3.post1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
